Flesh out html and css
Added most of the html and css needed for completion.
Still have to add buttons. Had trouble with the alt-image
grey box in the hero, hoping I can find a better solution
than what I have now, which is basically
a div with an image inside of it. Not sure how to get
rid of the alt-image logo next to alt text.